+++ view/arbitre/match-poule-view.php
@@ -10,136 +10,6 @@
         <link rel="icon" href="../../img/esporter-icon.png">
     </header>
     <body>
-        <?php
-        
-            ## Importation des fichiers ##
-            session_start();
-            require_once(realpath(dirname(__FILE__) . '/../../class/header.php'));
-            require_once(realpath(dirname(__FILE__) . '/../../SQL.php'));  
-            require_once(realpath(dirname(__FILE__) . '/../admin/bracket.php'));
-            
-            
-            $afficherBoutonFermerResultat  = False;
-            
-            $header = new header(2);
-            if ($_SESSION['role'] == "arbitre") {
-                echo $header->customizeHeader($_SESSION['role']);
-            } else {
-                header('Location: ../../acces-refuse.php');
-            }
-
-            //Sql
-            
-            $sql = new requeteSQL();
-
-          
-            $bracket = new bracket();
-
-            //Id
-           $idTournoi = $_GET["id_tournoi"];
-            $idJeu = $_GET["id_jeu"];
-
-            //Nom tournoi par id_tournoi
-            $reqNomTournoi = $sql -> getTournoiNomByIdTournoi($idTournoi) -> fetch()[0];
-            $poulesTermines = false;
-            $pouleFinaleCreer = false;
-            $idPouleFinale = 0;
-            
-             
-
-            //Poule par id_tournoi
-            $reqPoule = $sql->getPouleIdTournoi($idTournoi);
-
-            //Valider les resultats
-            if (isset($_POST["valider"])){
-                if (!$bracket->pouleTerminer(array_slice($reqPoule, 0, 4)))
-                        if (isset($_POST["match1"])){
-                            $req = $sql->setGagnantRencontre($_POST["idMatch1"], $_POST["match1"]);
-                        }
-
-                        if (isset($_POST["match2"])){
-                            $req = $sql->setGagnantRencontre($_POST["idMatch2"], $_POST["match2"]);
-                        }
-
-                        if (isset($_POST["match3"])){
-                            $req = $sql->setGagnantRencontre($_POST["idMatch3"], $_POST["match3"]);
-                        }
-
-                        if (isset($_POST["match4"])){
-                            $req = $sql->setGagnantRencontre($_POST["idMatch4"], $_POST["match4"]);
-                        }
-
-                        if (isset($_POST["match5"])){
-                            $req = $sql->setGagnantRencontre($_POST["idMatch5"], $_POST["match5"]);
-                        }
-
-                        if (isset($_POST["match6"])){
-                            $req = $sql->setGagnantRencontre($_POST["idMatch6"], $_POST["match6"]);
-                        }
-
-                    //Verifier que tout les matchs sont terminer
-                    $poulesTermines = true;
-                    
-                    
-                    if ($bracket->pouleTerminer(array_slice($reqPoule, 0, 4))){
-                        echo "Toute les poules sont terminées";
-                        $idPouleFinale = $bracket->genererPouleFinale($idTournoi,$sql->getIdJeu($idJeu),array_slice($reqPoule, 0, 4));
-                        
-                    }else{
-                        echo "Toute les poules ne sont pas terminées";
-                    }
-                        
-       
-                }else if ($pouleFinaleCreer){// Si poule terminer
-
-                //                                                                                                                                  ;
-                    print_r ($reqPoule);
-                    
-                    
-                    
-                  
-                 }
-
-                 if (isset($_POST["valider_finale"])){
-                    
-                    if (isset($_POST["match1"])){
-                        $req = $sql->setGagnantRencontreFinale($_POST["idMatch1"], $_POST["match1"]);
-                    }
-
-                    if (isset($_POST["match2"])){
-                        $req = $sql->setGagnantRencontreFinale($_POST["idMatch2"], $_POST["match2"]);
-                    }
-
-                    if (isset($_POST["match3"])){
-                        $req = $sql->setGagnantRencontreFinale($_POST["idMatch3"], $_POST["match3"]);
-                    }
-
-                    if (isset($_POST["match4"])){
-                        $req = $sql->setGagnantRencontreFinale($_POST["idMatch4"], $_POST["match4"]);
-                    }
-
-                    if (isset($_POST["match5"])){
-                        $req = $sql->setGagnantRencontreFinale($_POST["idMatch5"], $_POST["match5"]);
-                    }
-
-                    if (isset($_POST["match6"])){
-                        $req = $sql->setGagnantRencontreFinale($_POST["idMatch6"], $_POST["match6"]);
-                    }
-                    $idPouleFinale = $sql->getIDPouleFinale($idTournoi,$sql->getIdJeu($idJeu));
-                    
-                    if ($sql->pouleFinaleTerminer($idPouleFinale)){
-                    $sql->terminerTournoi($idTournoi);
-                        
-                        $bracket->updateClassementGeneral($idPouleFinale,$idTournoi);
-                    }else{
-                        
-                    }
-                 }
-
-                 $reqPoule = $sql->getPouleByIdTournoi($idTournoi);
-                
-           
-        ?>
 
         <main class="main-listes">
             <section class="main-listes-container">
